Move main function back into the scripts/ dir
